CUI UPDATE
Changes to be committed:
	modified:   src/domain/ArticleAdministration.java
	modified:   src/domain/LogAdministration.java
	modified:   src/domain/Shop.java
	modified:   src/domain/UserAdministration.java
	new file:   src/domain/exceptions/ArticleNotFoundException.java
	modified:   src/persistence/FilePersistenceManager.java
	modified:   src/persistence/PersistenceManager.java
	modified:   src/ui/CUI.java
	modified:   src/valueObject/Article.java
	modified:   src/valueObject/Invoice.java
	modified:   src/valueObject/ShoppingCart.java
	modified:   src/valueObject/User.java

+++ src/valueObject/Article.java
@@ -63,7 +63,7 @@ public class Article {
 
     public void setArticleNr(int articleNr)     { this.articleNr = articleNr; }
 
-    public void setPrice(float price)           { this.price = price; }
+    public void setPrice(double price)           { this.price = price; }
 
     public void setStock(int stock)             { logStock(stock - this.stock); this.stock = stock; }
